Make account_signup into its own view module
